fix no messages path

+++ nodestream/pipeline/extractors/queues/sqs.py
@@ -66,12 +66,15 @@ class SQSQueueConnector(QueueConnector, alias="sqs"):
     async def get_next_messsage_batch(self):
         loop = asyncio.get_running_loop()
         msgs = await loop.run_in_executor(None, self.get_message_batch)
-        if len(msgs["Messages"]) == 0:
+        if msgs.get("Messages") is None:
             return None
         return self.process_messages(msgs)
 
     def process_messages(self, msgs):
-        for message in msgs["Messages"]:
+        messages = msgs.get("Messages")
+        if messages is None:
+            return None
+        for message in messages:
             yield message["Body"]
         if self.delete_after_read:
             msgs.delete()
